app/src/main/java/com/beettechnologies/agreena/common/data/model/ApiResponse.kt
package com.beettechnologies.agreena.common.data.model

import com.google.gson.Gson
import retrofit2.Response

sealed class ApiResponse<T> {

    companion object {

        fun <T> create(error: Throwable): ApiErrorResponse<T> {
            return ApiErrorResponse(error.message ?: "unknown error")
        }

        fun <T> create(response: Response<T>): ApiResponse<T> {
            return if (response.isSuccessful) {
                val body = response.body()
                if (body == null || response.code() == 204) {
                    ApiEmptyResponse(code = response.code(), message = response.message())
                } else {
                    ApiSuccessResponse(body = body, code = response.code(), message = response.message())
                }
            } else {
                val message = response.errorBody()?.string()
                val errorMessage = if (message.isNullOrEmpty()) {
                    response.message()
                } else {
                    val error = Gson().fromJson(message, Error::class.java)
                    val errorList = error.detail.split(":")
                    if (errorList.size > 1) {
                        errorList[1]
                    } else {
                        error.detail
                    }
                }
                ApiErrorResponse(errorMessage)
            }
        }
    }

    data class Error(val detail: String, val code: String)
    class ApiEmptyResponse<T>(val code: Int, val message: String) : ApiResponse<T>()
    data class ApiSuccessResponse<T>(val body: T, val code: Int, val message: String) : ApiResponse<T>()
    data class ApiErrorResponse<T>(val errorMessage: String) : ApiResponse<T>()
}